+++ cli/cli-debug/src/test/java/io/telicent/smart/cache/cli/commands/debug/DockerTestDebugCliMutualTlsKafka.java
@@ -18,35 +18,38 @@ package io.telicent.smart.cache.cli.commands.debug;
 import ch.qos.logback.classic.Level;
 import io.telicent.smart.cache.cli.commands.AbstractCommandTests;
 import io.telicent.smart.cache.cli.commands.SmartCacheCommandTester;
-import io.telicent.smart.cache.live.LiveReporter;
 import io.telicent.smart.cache.sources.Header;
-import io.telicent.smart.cache.sources.kafka.FlakyKafkaTest;
-import io.telicent.smart.cache.sources.kafka.KafkaEventSource;
-import io.telicent.smart.cache.sources.kafka.KafkaTestCluster;
-import io.telicent.smart.cache.sources.kafka.SecureKafkaTestCluster;
+import io.telicent.smart.cache.sources.kafka.*;
 import io.telicent.smart.cache.sources.kafka.sinks.KafkaSink;
 import io.telicent.smart.cache.sources.memory.SimpleEvent;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.jena.riot.WebContent;
-import org.apache.jena.riot.web.HttpNames;
 import org.apache.kafka.common.serialization.StringSerializer;
 import org.testng.Assert;
+import org.testng.SkipException;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-import java.util.*;
+import java.io.File;
+import java.util.Collection;
+import java.util.Collections;
 
 import static io.telicent.smart.cache.cli.commands.debug.TestLogUtil.enableSpecificLogging;
 
-public class DockerTestDebugCliSecureKafka extends AbstractCommandTests {
+public class DockerTestDebugCliMutualTlsKafka extends AbstractCommandTests {
 
-    private final SecureKafkaTestCluster kafka = new SecureKafkaTestCluster();
+    private static final String CLIENT_PROPERTIES_FILE = new File("test-certs/client.properties").getAbsolutePath();
+    private final MutualTlsKafkaTestCluster kafka = new MutualTlsKafkaTestCluster();
 
     @BeforeClass
     @Override
     public void setup() {
+        if (StringUtils.contains(System.getProperty("os.name"), "Windows")) {
+            throw new SkipException(
+                    "These tests cannot run on Windows because the SSL certificates generator script assumes a Posix compatible OS");
+        }
+
         // Uncomment for easier debugging in IDE
         //SmartCacheCommandTester.TEE_TO_ORIGINAL_STREAMS = true;
         super.setup();
@@ -56,12 +59,10 @@ public class DockerTestDebugCliSecureKafka extends AbstractCommandTests {
 
     private void setupLogging() {
         enableSpecificLogging(KafkaEventSource.class, Level.DEBUG);
-        enableSpecificLogging(LiveReporter.class, Level.INFO);
     }
 
     private void teardownLogging() {
         enableSpecificLogging(KafkaEventSource.class, Level.OFF);
-        enableSpecificLogging(LiveReporter.class, Level.OFF);
     }
 
 
@@ -70,7 +71,6 @@ public class DockerTestDebugCliSecureKafka extends AbstractCommandTests {
     public void testCleanup() throws InterruptedException {
         super.testCleanup();
         this.kafka.resetTestTopic();
-        this.kafka.resetTopic(LiveReporter.DEFAULT_LIVE_TOPIC);
 
         // Occasionally we can get random test errors because the test topic(s) don't get recreated in time which can
         // lead to incomplete test data being generated in the subsequent test.  A short sleep makes that unlikely to
@@ -96,9 +96,7 @@ public class DockerTestDebugCliSecureKafka extends AbstractCommandTests {
                                                        .valueSerializer(StringSerializer.class)
                                                        .bootstrapServers(this.kafka.getBootstrapServers())
                                                        .topic(KafkaTestCluster.DEFAULT_TOPIC)
-                                                       .producerConfig(this.kafka.getClientProperties(
-                                                               this.kafka.getAdminUsername(),
-                                                               this.kafka.getAdminPassword()))
+                                                       .producerConfig(this.kafka.getClientProperties())
                                                        .lingerMs(5)
                                                        .build()) {
             for (int i = 1; i <= 1_000; i++) {
@@ -108,88 +106,63 @@ public class DockerTestDebugCliSecureKafka extends AbstractCommandTests {
         }
     }
 
-    private void runDumpCommand(String command, String... extraArgs) {
-        List<String> args = new ArrayList<>(Arrays.asList(command,
-                                                          "--bootstrap-servers",
-                                                          this.kafka.getBootstrapServers(),
-                                                          "--topic",
-                                                          KafkaTestCluster.DEFAULT_TOPIC,
-                                                          "--kafka-user",
-                                                          this.kafka.getAdminUsername(),
-                                                          "--kafka-password",
-                                                          this.kafka.getAdminPassword(),
-                                                          "--max-stalls",
-                                                          "1",
-                                                          "--poll-timeout",
-                                                          "3",
-                                                          "--read-policy",
-                                                          "BEGINNING",
-                                                          "--no-health-probes",
-                                                          "--no-live-reporter"));
-        if (extraArgs != null && extraArgs.length > 0) {
-            args.addAll(Arrays.asList(extraArgs));
-        }
-        DebugCli.main(args.toArray(new String[0]));
+
+    private void runDumpCommand(String dump) {
+        DebugCli.main(new String[] {
+                dump,
+                "--bootstrap-servers",
+                this.kafka.getBootstrapServers(),
+                "--topic",
+                KafkaTestCluster.DEFAULT_TOPIC,
+                "--kafka-properties",
+                CLIENT_PROPERTIES_FILE,
+                "--max-stalls",
+                "1",
+                "--poll-timeout",
+                "5",
+                "--read-policy",
+                "BEGINNING",
+                "--no-live-reporter",
+                "--no-health-probes"
+        });
     }
 
     @Test
-    public void debug_dump_01() {
+    public void givenEmptyTopic_whenDumpingEvents_thenNothingDumped() {
+        // Given and When
         runDumpCommand("dump");
 
+        // Then
         AbstractDockerDebugCliTests.verifyDumpCommandUsed();
-
         String stdErr = SmartCacheCommandTester.getLastStdErr();
         Assert.assertTrue(StringUtils.contains(stdErr, "Currently no new events available"));
     }
 
     @Test(retryAnalyzer = FlakyKafkaTest.class)
-    public void debug_dump_02() {
-        // Add some sample events to Kafka
+    public void givenNonEmptyTopic_whenDumpingEvents_thenEventsAreDumped() {
+        // Given
         generateKafkaEvents("Event %,d");
 
+        // When
         runDumpCommand("dump");
 
+        // Then
         AbstractDockerDebugCliTests.verifyDumpCommandUsed();
-
         AbstractDockerDebugCliTests.verifyEvents("Event %,d");
         String stdErr = SmartCacheCommandTester.getLastStdErr();
         Assert.assertTrue(StringUtils.contains(stdErr, "Currently no new events available"));
     }
 
     @Test(retryAnalyzer = FlakyKafkaTest.class)
-    public void debug_rdf_dump_01() {
-        // Add some sample events to Kafka
+    public void givenNonEmptyTopic_whenDumpingRdfEvents_thenEventsAreDumped() {
+        // Given
         generateKafkaEvents("<http://subject> <http://predicate> \"%d\" .");
 
+        // When
         runDumpCommand("rdf-dump");
 
+        // Then
         AbstractDockerDebugCliTests.verifyRdfDumpCommandUsed();
-
-        AbstractDockerDebugCliTests.verifyEvents("\"%d\"");
-    }
-
-    @Test(retryAnalyzer = FlakyKafkaTest.class)
-    public void debug_rdf_dump_02() {
-        // Add some sample events to Kafka
-        generateKafkaEvents("<http://subject> <http://predicate> \"%d\" .");
-
-        runDumpCommand("rdf-dump", "--output-language", "no-such-language");
-
-        AbstractDockerDebugCliTests.verifyRdfDumpCommandUsed();
-
-        AbstractDockerDebugCliTests.verifyEvents("\"%d\"");
-    }
-
-    @Test(retryAnalyzer = FlakyKafkaTest.class)
-    public void debug_rdf_dump_03() {
-        // Add some sample events to Kafka
-        generateKafkaEvents(List.of(new Header(HttpNames.hContentType, WebContent.ctPatch.getContentTypeStr())),
-                            "A <http://subject> <http://predicate> \"%d\" .");
-
-        runDumpCommand("rdf-dump");
-
-        AbstractDockerDebugCliTests.verifyRdfDumpCommandUsed();
-
         AbstractDockerDebugCliTests.verifyEvents("\"%d\"");
     }
 }