+++ event-sources/event-source-kafka/src/test/java/io/telicent/smart/cache/sources/kafka/DockerTestKafkaEventSource.java
@@ -60,7 +60,7 @@ public class DockerTestKafkaEventSource {
     @BeforeClass
     public void setup() {
         Utils.logTestClassStarted(this.getClass());
-        this.kafka = new KafkaTestCluster();
+        this.kafka = new BasicKafkaTestCluster();
         this.kafka.setup();
     }