+++ event-sources/event-source-kafka/src/test/java/io/telicent/smart/cache/sources/kafka/SecureKafkaTestCluster.java
@@ -32,9 +32,10 @@ import java.util.Map;
 import java.util.Properties;
 
 /**
- * A variation upon {@link KafkaTestCluster} that enables Kafka's user authentication features
+ * A variation upon {@link KafkaTestCluster} that enables Kafka's user authentication features with SASL PLAINTEXT
+ * authentication
  */
-public class SecureKafkaTestCluster extends KafkaTestCluster {
+public class SecureKafkaTestCluster extends BasicKafkaTestCluster {
 
     public static final String DEFAULT_ADMIN_USERNAME = "admin";
 
@@ -109,7 +110,7 @@ public class SecureKafkaTestCluster extends KafkaTestCluster {
     protected KafkaContainer createKafkaContainer() {
         // Heavily inspired by https://github.com/michelin/ns4kafka/blob/master/src/test/java/com/michelin/ns4kafka/integration/AbstractIntegrationTest.java
         //@formatter:off
-        return new KafkaContainer(DockerImageName.parse("confluentinc/cp-kafka:7.3.0"))
+        return new KafkaContainer(DockerImageName.parse("confluentinc/cp-kafka:7.7.1"))
                 .withKraft()
                 .withEnv("KAFKA_LISTENER_SECURITY_PROTOCOL_MAP", "PLAINTEXT:SASL_PLAINTEXT,BROKER:SASL_PLAINTEXT")
                 .withEnv("KAFKA_INTER_BROKER_LISTENER_NAME", "BROKER")